teachmeskills/test_task_fast_api_aiogram/fast_api_server/pyproject.toml
[tool.poetry]
name = "fastapi test task"
version = "0.1.0"
description = ""
authors = ["jahoroshi4y <jahoroshi4y@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.114.2"
alembic = "^1.13.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.34"}
passlib = "^1.7.4"
aiosqlite = "^0.20.0"
email-validator = "^2.2.0"
python-jose = "^3.3.0"
python-multipart = "^0.0.9"
uvicorn = "^0.30.6"
pycryptodome = "^3.20.0"
fastapi-limiter = "^0.1.6"
jinja2 = "^3.1.4"
redis = "^5.0.8"
setuptools = "^75.0.0"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
